feat: enhance category management with user-specific data handling; update user ID integration in add, edit, delete, and list operations

+++ backend/modules/Categories/deleteCategory/Repository.ts
@@ -1,8 +1,8 @@
 import categoryModel from '../../../models/Category';
 
 class categoryRepository {
-	async deleteCategoryById(id: string) {
-		return categoryModel.findOneAndDelete({ id });
+	async deleteCategoryById(id: string, userID: string) {
+		return categoryModel.findOneAndDelete({ id, userID });
 	}
 }
